[TASK] made elements typings
