+++ src/models/Badge.d.ts
@@ -1,0 +1,9 @@
+/**
+ * The badge of an user
+ */
+export interface BadgeInterface {
+  /**
+   * The type of the badge
+   */
+  type: string;
+}
