+++ src/models/Descriptor.d.ts
@@ -34,18 +34,3 @@ export interface DescriptorInterface {
    */
   choice_selections: ChoiceSelectionInterface[];
 }
-
-/**
- * Implements some basic methods for interaction
- *
- * @class Descriptor
- */
-class Descriptor {
-  private descriptorData: DescriptorInterface;
-
-  constructor(data: DescriptorInterface) {
-    this.descriptorData = data;
-  }
-}
-
-export default Descriptor;
