+++ src/models/Hashes.d.ts
