+++ src/models/Position.d.ts
@@ -15,18 +15,3 @@ export interface PositionInterface {
    */
   lon: number;
 }
-
-/**
- * Implements some basic methods for interaction
- *
- * @class Position
- */
-class Position {
-  private positionData: PositionInterface;
-
-  constructor(data: PositionInterface) {
-    this.positionData = data;
-  }
-}
-
-export default Position;
