+++ src/models/SexualOrientation.d.ts
@@ -1,0 +1,13 @@
+/**
+ * The sexual orientation of an user
+ */
+export interface SexualOrientationInterface {
+  /**
+   * The ID of the sexual orientation
+   */
+  id: string;
+  /**
+   * The name of the sexual orientation
+   */
+  name: string;
+}
