+++ src/models/SizedImage.d.ts
