+++ src/models/SpotifyThemeTrack.d.ts
@@ -63,18 +63,3 @@ export interface SpotifyTrackInterface {
    */
   uri: string;
 }
-
-/**
- * Implements some basic methods for interaction
- *
- * @class SpotifyTrack
- */
-class SpotifyTrack {
-  private trackData: SpotifyTrackInterface;
-
-  constructor(data: SpotifyTrackInterface) {
-    this.trackData = data;
-  }
-}
-
-export default SpotifyTrack;
