+++ src/models/Teaser.d.ts
@@ -1,0 +1,13 @@
+/**
+ * The teaser of an user
+ */
+export interface TeaserInterface {
+  /**
+   * The type of an teaser
+   */
+  type: string;
+  /**
+   * The value string of an teaser
+   */
+  string: string;
+}
